+++ packages/theme/src/theme/fontSizes.ts
@@ -1,0 +1,1 @@
+export const fontSizes: { [key: string]: number | string } = {}