+++ packages/theme/src/theme/fontWeights.ts
@@ -1,0 +1,1 @@
+export const fontWeights: { [key: string]: number | string } = {}