+++ packages/theme/src/theme/lineHeights.ts
@@ -1,0 +1,1 @@
+export const lineHeights: { [key: string]: number } = {}